+++ inc/ti/varr.inline.h
@@ -16,12 +16,7 @@
 
 static inline _Bool ti_varr_may_have_things(ti_varr_t * varr)
 {
-    return varr->flags & TI_VARR_FLAG_MHT;
-}
-
-static inline _Bool ti_varr_may_gen_ids(ti_varr_t * varr)
-{
-    return varr->flags & (TI_VARR_FLAG_MHT|TI_VARR_FLAG_MHR);
+    return varr->flags & TI_VFLAG_MHT;
 }
 
 static inline _Bool ti_varr_is_list(ti_varr_t * varr)
@@ -62,7 +57,7 @@ static inline void * ti_varr_key(ti_varr_t * varr)
 
 static inline void ti_varr_set_may_flags(ti_varr_t * to, ti_varr_t * from)
 {
-    to->flags |= from->flags & (TI_VARR_FLAG_MHT|TI_VARR_FLAG_MHR);
+    to->flags |= ti_val_may_flags(from);
 }
 
 static inline _Bool ti_varr_is_stored(ti_varr_t * varr)