+++ src/ti/vset.c
@@ -101,7 +101,7 @@ int ti_vset_to_list(ti_vset_t ** vsetaddr)
 
     list->ref = 1;
     list->tp = TI_VAL_ARR;
-    list->flags = list->vec->n ? TI_VARR_FLAG_MHT : 0;
+    list->flags = list->vec->n ? TI_VFLAG_MHT : 0;
     list->parent = NULL;
 
     ti_val_unsafe_drop((ti_val_t *) *vsetaddr);
@@ -126,7 +126,7 @@ int ti_vset_to_tuple(ti_vset_t ** vsetaddr)
 
     tuple->ref = 1;
     tuple->tp = TI_VAL_ARR;
-    tuple->flags = TI_VARR_FLAG_TUPLE | (tuple->vec->n ? TI_VARR_FLAG_MHT : 0);
+    tuple->flags = TI_VARR_FLAG_TUPLE | (tuple->vec->n ? TI_VFLAG_MHT : 0);
 
     ti_val_unsafe_drop((ti_val_t *) *vsetaddr);
     *vsetaddr = (ti_vset_t *) tuple;